+++ src/replacer/index.ts
@@ -2,18 +2,20 @@ import { transformer } from "src/processor/ast"
 
 export interface ReplacerOptions {
     callee?: string | string[],
+    nestingCallee?: string | string[],
     separator?: string | false,
     debug?: boolean
 }
 
 /**
  * Transforms the content before Tailwind scans/extracting its classes.
- * @param {ReplacerOptions} [options = {callee="twg", separator=":", debug=true}] callee, separator, debug. See [docs](https://github.com/hoangnhan2ka3/twg?tab=readme-ov-content#replacer-options).
+ * @param {ReplacerOptions} [options = {callee="twg", nestingCallee=undefined, separator=":", debug=true}] callee, nestingCallee, separator, debug. See [docs](https://github.com/hoangnhan2ka3/twg?tab=readme-ov-content#replacer-options).
  * @param {string} content The content already provided by `content.files` in `tailwind.config`.
  * @returns string
  */
 export function replacer({
     callee = "twg",
+    nestingCallee,
     separator = ":",
     debug = true
 }: ReplacerOptions = {}) {
@@ -25,7 +27,7 @@ export function replacer({
         }
 
         try {
-            return transformer(content, { callee, separator, debug })
+            return transformer(content, { callee, nestingCallee, separator, debug })
         } catch { return content }
     }
 }